requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
bcrypt==4.0.1
tzdata>=2024.1
Pillow>=10.3.0
pillow-heif>=0.16.0
gunicorn>=20.1.0
